services/users/requirements.txt
# Flask
Flask==0.12.2

# Database (Postgres)
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.2
flask-migrate==2.1.1

# Testing
Flask-Testing==0.6.2
coverage==4.4.2
flake8==3.5.0
flask-debugtoolbar==0.10.1

# Web Server
gunicorn==19.7.1
flask-cors==3.0.3
